+++ History.md
@@ -1,3 +1,13 @@
+# 1.5.5-SNAPSHOT / ????-??-??
+
+## New and Noteworthy
+
+## Enhancements
+
+## Fixed Issues
+
+## External Contributions
+
 # 1.5.4 / 2023-05-27
 
 ## Enhancements